# services/api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "این سرویس به عنوان درگاه اصلی برای دسترسی به میکروسرویس‌های امانت و کاربران عمل می‌کند.",
        "title": "Books \u0026 Trust API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/login": {
            "post": {
                "description": "Accepts user credentials, validates them against the user service via gRPC, and returns a JWT access token upon successful authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Authenticate and login user",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http_user_handler.LoginUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ \"message\": \"user\t\t\t\t\t\t\t\t\t\t\t\t\t\tsuccessfully\tlogged\tin\", \"data\": { \"user_id\": \"string\", \"token\": \"string\", \"token_type\": \"string\" } }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{ \"code\": \"BAD_REQUEST\", \"message\": \"string\", \"details\": \"invalid\t\tuser\t\t\tlogin\trequest\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "{ \"code\": \"UNAUTHENTICATED\", \"message\": \"string\", \"details\": \"invalid\tusername\t\tor\t\tpassword\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{ \"code\": \"INTERNAL_SERVER_ERROR\", \"message\": \"string\", \"details\": \"\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Accepts user registration details and registers a new account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http_user_handler.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "{ \"message\": \"user\tsuccessfully\tregistered\", \"data\": { \"user_id\": \"string\", \"email\": \"string\", \"created_at\": \"string\" } }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{ \"code\": \"BAD_REQUEST\", \"message\": \"string\", \"details\": \"string\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{ \"code\": \"INTERNAL_SERVER_ERROR\", \"message\": \"string\", \"details\": \"\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/update": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing user's information via gRPC user service.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user details",
                "parameters": [
                    {
                        "description": "User update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http_user_handler.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ \"message\": \"user\t\t\t\t\t\t\t\t\t\t\t\t\tsuccessfully\tupdated\", \"data\": { \"user_id\": \"string\", \"name\": \"string\", \"email\": \"string\", \"username\": \"string\" } }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{ \"code\": \"BAD_REQUEST\", \"message\": \"string\", \"details\": \"invalid\tuser\t\t\tupdate\trequest\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{ \"code\": \"INTERNAL_SERVER_ERROR\", \"message\": \"string\", \"details\": \"\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches user details from the user microservice using their unique ID provided in the URL path.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ \"message\": \"user\t\t\t\t\t\t\t\t\t\t\t\t\tsuccessfully\tfetched\", \"data\": { \"user_id\": \"string\", \"name\": \"string\", \"email\": \"string\", \"username\": \"string\" } }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{ \"code\": \"BAD_REQUEST\", \"message\": \"string\", \"details\": \"invalid\tget\t\t\t\tuser\trequest\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{ \"code\": \"INTERNAL_SERVER_ERROR\", \"message\": \"string\", \"details\": \"\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a user account from the system using their unique ID provided in the URL path. Only the authenticated user themselves can delete their own account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete user account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ \"message\": \"user\t\t\t\t\t\t\t\t\t\t\t\t\tsuccessfully\tdeleted\", \"data\": null }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{ \"code\": \"BAD_REQUEST\", \"message\": \"string\", \"details\": \"invalid\tdelete\t\t\tuser\trequest\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "{ \"code\": \"FORBIDDEN\", \"message\": \"string\", \"details\": \"you\t\tdo\t\t\t\tnot\t\thave\tpermission\tto\tdelete\tthis\tuser\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{ \"code\": \"INTERNAL_SERVER_ERROR\", \"message\": \"string\", \"details\": \"\" }",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_handler_http_user_handler.LoginUserRequest": {
            "type": "object",
            "required": [
                "credential",
                "password"
            ],
            "properties": {
                "credential": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_handler_http_user_handler.RegisterUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
        "internal_handler_http_user_handler.UpdateUserRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "تایپ کنید: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}